Revert type casting as it affected many other parsers
