+++ golem-rib/src/parser/mod.rs
@@ -19,3 +19,4 @@ mod select_index;
 mod sequence;
 mod tuple;
 mod type_name;
+pub(crate) mod type_binding;
